requirements.txt

Flask==2.0.1
SQLAlchemy==1.4.22
Pillow==8.3.1
Celery==5.1.2
redis==3.5.3
requests==2.26.0
